myapp/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
